+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - feature: new module `echo.send_dialog`, to send the text using `dialog GUI`.
+- [core] feature: now will only send the `error stack trace` when `exceptions in command execution`, if the command source is admin.